some minor correction to port variable & delete controller

+++ backend/controllers/product.controller.js
@@ -55,6 +55,11 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
   // console.log("id: ", id);
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
 
   try {
     await Product.findByIdAndDelete(id);
